Add GitHub Action for automated testing and building
A new GitHub Action 'build-and-test.yml' has been added to automate the process of testing and building on each push and pull request. The Action is configured to run on multiple OS versions and uses .NET Core 8.0.100 for the operations. All *.cs and *.csproj files in the main branch will trigger this Action.

In addition to this, various test methods from ConfigFileExtensionsTests.cs have been removed as part of a broader refactor, since their functionality is covered elsewhere or is no longer required.